* fix double click style for CTextEdit 	* UIDescription: add variable substitution for all attributes when creating views (thanks to Wolfram Franke for the inital version)

+++ vstgui/lib/cview.h
@@ -127,7 +127,7 @@ extern UTF8StringPtr kPerthousandSymbol;	///< per mille sign
 extern IdStringPtr kMsgViewSizeChanged;		///< Message send to parent that the size of the view has changed
 
 //-----------------------------------------------------------------------------
-typedef uint32_t CViewAttributeID;
+typedef size_t CViewAttributeID;
 //-----------------------------------------------------------------------------
 // Attributes
 //		all attributes where the first letter is lowercase are reserved for the vstgui lib
